feat: feature switch for new deletion workflows

+++ src/main/java/de/caritas/cob/userservice/api/deleteworkflow/service/DeleteUsersRegisteredOnlyService.java
@@ -25,6 +25,8 @@ public class DeleteUsersRegisteredOnlyService {
   private final @NonNull DeleteUserAccountService deleteUserAccountService;
   private final @NonNull WorkflowErrorMailService workflowErrorMailService;
 
+  @Value("${user.registeredonly.deleteWorkflow.enabled}")
+  private boolean userRegisteredOnlyDeleteWorkflowEnabled;
   @Value("${user.registeredonly.deleteWorkflow.check.days}")
   private int userRegisteredOnlyDeleteWorkflowCheckDays;
 
@@ -32,6 +34,10 @@ public class DeleteUsersRegisteredOnlyService {
    * Deletes all askers without running sessions.
    */
   public void deleteUserAccounts() {
+
+    if (!userRegisteredOnlyDeleteWorkflowEnabled) {
+      return;
+    }
 
     var workflowErrors = deleteAskersAndCollectPossibleErrors();
